Update Labelgenerator_1.py
The labels are now not printed in a single row. Instead after every 20 labels, the next column is chosen. This can be done until column 'Z' (~520 labels possible).

The existence of a target worksheet is now checked.
If it exists, a new sheet is generated, if it does not exist it is generated.

Sheets are now labelled with date.